Creates searchcraft-input-form to replace auto-search-form and base-search-form. re:SC-554
Combines the functionality of base-search-form, auto-search-form, and searchcraft-input, into a new single component, searchcraft-input-form. searchcraft-input-form is capable of rendering all the different input layouts and functionality previously that was previously split across those three components.

This is to prepare for the popover form component, that also needs to use an instance of `searchcraft-input-form`. We needed a single component so that we don't needlessly replicate/repeat functionality across multiple components.