Updates to core initialization behavior, package export settings, build scripts re:SC-731
Makes changes to how the sdk is initialized by the consuming application, so that the SearchcraftCore object is initialized independently from the components. It also makes changes to the stencil build settings so that the unnecessary build targets are gone, and so that the generated components module is the top-level export of the package.

Moves searchcraftStore to core & updates build settings re:SC-731

Update searchcraft initialization, WIP

Core functionality WIP

Update core init behavior